gn2bp: use Android libgtest_prod_headers
This requires aosp/2366787 to work.
Unfortunately, Chromium does not add a dependency on gtest, so
builtin_module cannot be used.
Using the gtest prod headers from Android is fine for //base. The only
thing that's used is gtest_prod.h, which includes one friend macro that
hasn't changed in 14 years.

Test: build cronet
Change-Id: I03b12ae698699527da7e69c9d4d50da1822fd5bc

Former-commit-id: 17d981bd3f2399b7dd8a04450e73b5b12f9632af