gn2bp: Turn off `allocator_shims` for musl only
* I've investigated turning off `allocator_shims` globally for all of Cronet but
this didn't sound like a good option. allocator_shims are used for security measures,
where sometimes malloc caller forget to check for NULL which happens on oom. allocator_shims would
instantly kill the app instead of depending on the caller to do the validity check.
* The solution here was to manually remove the files and add them to `android` and `glibc` only so allocator_shims
is never compiled for musl.
* Chromium does not support musl so there's no flag for `musl` with allocator_shims.

Test: mma
Change-Id: I2a63b14482a526d06a089647e6f61eea9f1691dc

Former-commit-id: 4895bb946d71203a0a455570458c5736292ca341