Make all @VisibleForTesting methods public.
In some cases the jarjar setup results in separating production and test
classes which live in the same package at compile time, leading to
crashes during runtime. Exposing the methods more widely has no other
practical effect - access to them is still disallowed for general
users.

The methods should not be migrated to @TestApi because
  1) they're part of the internal package which gets jarjared, and
  2) @TestApi is discouraged for Mainline code anyway

Test: m
Change-Id: I75b5c41449192040d70c3b2083350847840b9812

Former-commit-id: 5f93f38167c2019c3f0083572d94b87d05d049f0