[Cronet] Start cronetEngine with testFramework in CronetLoggerTest
The instance of CronetEngine being shutdown at teardown
is different from the one that was started in the tests.
This cl switches the cronetEngine being started to the same as the
one being shutdown.

This became an noticable because another test kept failing with
"Disk cache storage path already in use" as the first engine
sets the storage cache and never shuts down, hence not releasing
that storage location. Only one engine is allowed to use a
storage location at a time. This was not seen in chromium
because one test is run per JVM.

Test: atest NetHttpCoverageTests
Bug: 267353182
Change-Id: Ibbee46ed73d1ebcdef338339624183fa1e68b875

Former-commit-id: 417d4d831f88f75f335d7f2b906ffb0083d18592