cronet: add post-commit hook to generate .patch file
This post-commit hook generates (and amends) a .patch file if a given
commit touches a .gn or .gni file. It will only include changes to these
specific files (.gn or .gni) in the patch.

Note that this hook is does not run during an interactive rebase.

This change also includes an install script that creates a symlink to
the post-commit hook in .git/hooks for convenience.

Test: git commit.
Change-Id: I532794c6a400cf2e10e3c0e5122ec3df33636d6c

Former-commit-id: e442143340b4d7bd13aebbb147e30a00b36cad7f