gn2bp: add ProtocJavaSanitizer support
Cronet has added an upstream protobuf dependency to the java module.
This change adds sanitizer support for java protobuf dependencies. Due
to the nature of action modules, the libprotobuf dependency cannot
easily be propagated back to the cronet_aml_java modules, so its added
manually (though it would be great if we had a more generic solution to
passing dependencies around -- one day, I guess).

Note that this uses libprotobuf from the Android tree. Since there are
no ABI issues in java, and protobuf generally is both forwards and
backwards compatible, this should not pose an issue.

Test: update_results.sh
Change-Id: I816e1c0bd9dd0613239561d2db3fab1e7007da41

Former-commit-id: 802dd24c52ece414ad40fee0ed161d47a282788e