cronet import: move NetHttpTestsLibPreJarJar to external/cronet
Classes in NetHttpTestsLibPreJarJar newly depend on the JNI annotation
preprocessor (to parse the @NativeMethods annotations). Since the
annotation prepocessor soong module gets generated by gn2bp and is not
visible to anything outside of //external/cronet, it is easiest to just
move the NetHttpTestsLibPreJarJar module into Android.extras.bp.

This has the added benefit, that in the future, we can just run `mm` in
//external/cronet to test the import.

The newly redundant filegroups will be cleaned up in a follow up change.

Test: mm
Change-Id: I5be692c85a7934463061e0294118eba70fee31c5

Former-commit-id: 22548ec417d76cf9cf70f84ec1c8ccb175801fc4