Cherry-pick https://crrev.com/c/4667153
Original commit message:
Revamp the Cronet telemetry manifest flag

Before:

<meta-data android:name="org.chromium.net.EnableCronetTelemetry" android:value="true" />

After:

<service android:name="android.net.http.MetaDataHolder">
	<meta-data android:name="net.android.http.EnableTelemetry" android:value="true" />
</service>

The rationale is as follows:

- The meta-data entries are moved to a "meta-data holder service" as application-level meta-data tends to cause scalability issues with PackageManager queries.

- The prefix is changed from "org.chromium.net" to "android.net.http" to make the names look less jarring when using Cronet in AOSP (i.e. HttpEngine). This means the AOSP prefix is also used even when using the Chromium variant of Cronet, but that seems more tolerable. We could also keep using the Chromium prefix in Chromium and adjust it downstream in AOSP, but that would add complexity for little benefit.

- The name "Cronet" is removed because that is redundant with the "android.net.http" prefix, and AOSP surfaces are not supposed to use the term "Cronet".

This change is safe from a backwards compatibility perspective because no apps use this flag today, except for one (AGSA), which has been adjusted to specify both the old and new meta-data entries so that it will behave correctly regardless of which Cronet version it's using.

We take the opportunity to also simplify the test code - we can simply replace the entire meta-data object instead of "patching" it because Cronet owns the entire Bundle.

Bug: 284972285
Test: TH
Change-Id: Iccdc60b640e49d1b47473bfb8c95e5f0f4cf60aa

Former-commit-id: 8c40b44bf9ffa410fb5ef1c1c815b6bbd6e5d1f4