Cherry-pick https://crrev.com/c/4687421
Original commit message:
Make CronetSources aware of platform Cronet

Cronet from the platform is always loaded from the bootclassloader. This
makes it possible to directly implement CRONET_SOURCE_PLATFORM upstream,
without having to fork the code downstream: all we have to do is check
for the classloader of Cronet's implementation (similarly to what we do
for Google Play Services Cronet).

Test: TH
Bug: 284972285
Change-Id: I62c54fee57dde59de686d714557fe757b74646b8

Former-commit-id: 2beb0754629ca4d7c05c6f229fe5d77c82d405d0