Cronet Import: add JavaJniGenerator
JniGenerator actions are sometimes created twice, once for the java side and another for the C++ side.
I added a new class called `JavaJniGenerator` which is responsible for handling different side.

Note: Both side almost have the same behaviour, a better solution would be to refactor Sanitizer by moving
`type` and `is_testing` to base class `ActionSanitizer` and apply the changes to every sanitizer.

libcronet.so now builds and links

Change-Id: I104914265bd45808ebcc62d3185d661046dade67

Former-commit-id: 261cf74e3621e6649815257b365a0ac47f9f6918