Rename libcronet.so to prevent conflict when calling System.loadLibrary()
When calling System.loadLibrary(), ART includes a check to ensure the
native library has not already been loaded from a different class loader
context: "The library will be associated with class_loader. The JNI spec
says we can't load the same library into more than one class loader."

This error is encountered regardless of the conflicting libraries being
in different linker namespaces (which would later correctly load /
resolve symbols).

This is an issue specifically for HttpEngine which calls
System.loadLibrary("cronet.114.0.5735.84") just as upstream Cronet does.
If an app happens to statically link upstream Cronet with version
114.0.5735.84, it will crash when trying to use HttpEngine.

The immediate solution is to rename HttpEngine's libcronet.so to
libhttpengine.so.

Bug: 298674099
Test: TH
Change-Id: I64f48f5337caa36e53d33d5a300ee0ed25aa29ec

Former-commit-id: 4f7ebec594e3f79ab72eb7401ea4bbba62bf5441