Tweak Proto target generation
Looking at the desc_*.json file, we can see that the proto targets has the following structure:

cc_library_static -> proto_cc_action
                 |-> proto_h_action

The previous implemented completely removed the `cc_library_static` and bubbled up the `proto_cc_action` and `proto_h_action` upward the tree which worked fine before we hit the linker issue. This no longer work because a `proto_cc_action` can be included multiple times as part of the srcs of different target which led to `multiple definition of a symbol` error. The fix for this is to define a `cc_library_static` where the `proto_cc_action` can be only listed once as `srcs` then propagate that `cc_library_static` upward the tree correctly.

Change-Id: I33a20e84513853d7028d4b0188cf11cbcb890408

Former-commit-id: 9f54d57b4e72db21438989a1cdbdc59da0f0fe18