Checkout apihelpers from upstream-import
Apihelpers were deleted from upstream-staging, I'm not quite sure why they were deleted
in the first place, maybe because we don't ship apihelpers? This change introduces apihelpers
but they are not included in API surface that we ship and they can not be included by **mistake**
as adding any new API method requires updating a `text` file and get API council approval.

From the above, I can conclude that it is better to checkout those files to reduce the
import process merge conflicts and reduce the divergence further.

Change-Id: I656bebb7dc11e579898ccfe07817f82f6b791f82

Former-commit-id: aaa3c06a23b0d038462a8b928b2c1703fc4e3d97