Move disabled net tests to TEST_MAPPING
Those tests are only disabled in AOSP and not on Chromium and they
would cause divergence on import. In order to avoid that, I moved those
tests to TEST_MAPPING instead, the benefit of doing that is that the
flaky tests can still run on postsubmit while not blocking the
presubmits.

Added a new option flags to ChromiumHostDrivenTest to allow passing
exclude-filter, this was not necessary but had to be done because of
b/342355769. Once b/342355769 is fixed, we can use "exclude-filter"
directly.

Added the gtest_output to the test logs for easier debugging.

Bug: 304217500
Test: TH
Change-Id: Id2c03b7ed2e77fc835c2ee8f7114059b57069bff

Former-commit-id: dbf2dce56d6eed159b3dedc98cc9c5f067dac0c5