Cronet 124.0.6367.42: Remove unused file and library
accessibility_test_framework in chromium tree is updated from 2.1 to 4.0
but accessibility_test_framework in aosp is 2.0.
This makes the build fails but accessibility_test_framework is used only
from BaseActivityTestRule.java and this file is not used in aosp.
So this CL update gen_android_bp and gn_utils.py to ignore the file and
libray.

Test: Build
Bug: 345656003
Change-Id: I2fd65106741f36f7b00658d7bb6156f389f3d9ed

Former-commit-id: 81c4597176108f9506cefcdfe6361dd3b3f714e6