Add license script to generate licensing for AOSP
In order to comply with AOSP compliance & license tooling, we are now
going to generate the following files:

1. METADATA
2. MODULE_LICENSE_XYZ
3. LICENSE

for every README.chromium file, this is a requirment in AOSP.

I have added a script that generates such files along with plenty of
unit-tests and an end-to-end test that makes sure that Cronet's repo
is compliant for every import that we do, there are some caveats for
this implementation which are discussed below:

1. Since we are generating the compliance files in AOSP, we will not
know about any bad license until we decide to import that specific
version of Cronet to AOSP. I reckon that this will be very rare as
a bad license should affect Chromium before AOSP, there are some cases
where a third party library could have a GPL license for linux machine
but non-GPL license for Android, those cases would be only found out
during import and have to be fixed manually by removing the culprit
third party package.

2. We have to maintain a list of the known licenses, this allows us to
control which licenses are allowed. Licenses that are not mentioned
will not be allowed by default.

3. README.chromium files with weird formatting have to be handled
manually, the script allows for intermediate patching so the actual
files are not touched but the intermediate data structures can be
manipulated to fix this easily.

Change-Id: Ic06299f985740111cb4a4264fd23c6a40310e2dc

Former-commit-id: 0d39549dfaa0b15d0932ee286387be002e07c224