Break down Android.bp into project-specific Android.bps
This is the first step for compliance with the SBOM where each Android.bp
should live in its own separate directory.

Currently, genrules (except proto genrules) still live in the
top-level directory as moving them to package-level is considered
hard as we need to refactor a huge part of the code to rerun
the sanitizers with the rebased files.

There is no license packages yet, the next step would be to connect the
licensing python classes(aosp/3180001) with gn2bp. This means that gn2bp
will be aware if a specific directory contains license/metadata from
which it can create a `license` module in the Android.bp for full
integration with SBOM tooling.

boringssl is excluded from this process as the Android.bp living inside
boringssl is manually maintained by the Cronet team, although we still
do generate some boringssl modules, those are only used for testing
and never shipped, the actual modules that gets shipped lives in
third_party/boringssl.

Test: mma
Bug: 358541596
Change-Id: Ie56d6b45ba6eb57b6fa2b0b30afcbdc7e9bdff8d

Former-commit-id: d65204f845f27a822144b58f19bed87301574505