Generate license and package modules in generated Android.bp
This is the final step towards compliance with SBOM through the
generation of `license` and `package` modules to define which license
is being used for each Android.bp, the current approach is that a
`license` module will be generated if a README.chromium file exists
in the same directory as the Android.bp, this will work correctly for
most (all when some violations are fixed) third party packages, for
first party packages; this is fixed through the `package` module where
the `package module` will point to the nearest ancestor `license`
module. If a license module is not found then the top-level chromium
license is used which is handled manually in Android.extras.bp.

Refactored some python modules to allow reusability for gn2bp, this is
needed for gn2bp to know how to fill-in the license fields. The python
modules could use a refactor in terms of structuring but that will be
done later once I have figured out the appropriate way to structure
python packages in AOSP.

Bug: 358541596
Change-Id: I839b82c1cca8caca7e28d43e73684874ed13e1bc

Former-commit-id: caf27191fce391b8b7e07de761bf511f02a70b4e